requirements.txt
psutil==5.8.0
py-cpuinfo==7.0.0
WMI==1.4.9
